Refactored TCP/serial clients to use a common implementation.
These apps are identical except for the type of transport they use.
